Cleaning up application initialization
* Splitting out hooks into a different module
* Remove unused performance controller
* Replacing PecanAPI abstraction into a simple builder function
* Moving RootController up and out of create_main_app()
